// src/app/types/global.d.ts
export type TableData = {
    id: string
    oper: string
    date: string
    term: string
    store: string
    term2: string
    profTS: string
    status: string
    areaBuy: string
    myPrice: number
    propos: string
    doubleOk: string
    plusPropos: string
    priceAnPropos: string
    comission: string
    toPrice: number
    create: string
    okCreating: string
    client: string
}

export type UserModel = {
    login: string
    password: string
}